requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==4.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.0.0
django-storages==1.12.3
djangorestframework==3.12.4
dropbox==11.25.0
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
ply==3.11
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stone==3.2.1
toml==0.10.2
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
psycopg2-binary ==2.8.6
gunicorn==20.1.0
django-extra-fields==3.0.2
shortuuid===1.0.8
